c/scheduling: made allocation_units top level type
